requirements.txt
llama-index
llama-index-readers-file
python-dotenv
llama-index-llms-gemini
llama-index-embeddings-huggingface
pdfplumber
streamlit
llama-index-readers-web